Rename: 분양신청/분양공고 패키지 분리
- adopt 패키지 하위에 notice, application 패키지 분리
