--- MyohanMeeting/src/main/java/meet/myo/domain/adopt/AdoptApplication.java
@@ -1,29 +1,0 @@
-package meet.myo.domain.adopt;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import meet.myo.domain.BaseAuditingListener;
-import meet.myo.domain.Member;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AdoptApplication extends BaseAuditingListener {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "adopt_application_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    private String content;
-
-    @ManyToOne
-    @JoinColumn(name = "adopt_notice_id")
-    private AdoptNotice adoptNotice;
-
-}
