+++ MyohanMeeting/src/main/java/meet/myo/domain/adopt/notice/AdoptNoticeStatus.java
@@ -1,4 +1,4 @@
-package meet.myo.domain.adopt;
+package meet.myo.domain.adopt.notice;
 
 public enum AdoptNoticeStatus {
     ACCEPTING, COMPLETE, CANCELED
