+++ MyohanMeeting/src/main/java/meet/myo/repository/AdoptApplicationRepository.java
@@ -1,6 +1,6 @@
 package meet.myo.repository;
 
-import meet.myo.domain.adopt.AdoptApplication;
+import meet.myo.domain.adopt.application.AdoptApplication;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
