+++ MyohanMeeting/src/main/java/meet/myo/repository/AdoptNoticeRepository.java
@@ -1,6 +1,6 @@
 package meet.myo.repository;
 
-import meet.myo.domain.adopt.AdoptNotice;
+import meet.myo.domain.adopt.notice.AdoptNotice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
